requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
pyyaml==6.0.2
pytz==2024.1
